# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Amazon Bid Engine API
  description: |
    Amazon広告自動入札提案エンジンのREST API。

    ## 認証
    すべてのエンドポイント（/health を除く）では `X-API-Key` ヘッダーによる認証が必要です。

    ## Retool統合
    このAPIはRetool Cloudからの利用を想定しています。
    - Base URL: Cloud Runのデプロイ先URL
    - Authentication: Header Auth (`X-API-Key`)
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://your-service.a.run.app
    description: Production (Cloud Run)
  - url: http://localhost:8080
    description: Local Development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for authentication
  schemas:
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 総件数
        limit:
          type: integer
          description: 取得件数上限
        offset:
          type: integer
          description: オフセット
        hasMore:
          type: boolean
          description: 次のページがあるか

    # ネガティブキーワード候補
...
    AutoExactSuggestion:
      type: object
      properties:
        suggestion_id:
          type: string
          format: uuid
          description: 候補ID
        execution_id:
          type: string
          description: 実行ID
        profile_id:
          type: string
          description: Amazon AdsプロファイルID
        campaign_id_auto:
          type: string
          description: AUTOキャンペーンID
        ad_group_id_auto:
          type: string
          description: AUTO広告グループID
        campaign_id_manual_target:
          type: string
          description: ターゲットMANUALキャンペーンID
        ad_group_id_manual_target:
          type: string
          description: ターゲットMANUAL広告グループID
        asin:
          type: string
          description: ASIN
        search_term:
          type: string
          description: 検索語句
        intent_cluster_id:
          type: string
          description: インテントクラスタID
        intent_cluster_label:
          type: string
          description: インテントクラスタラベル
        match_type:
          type: string
          description: マッチタイプ
        lookback_days:
          type: integer
          description: 参照期間（日）
        impressions:
          type: integer
          description: インプレッション
        clicks:
          type: integer
          description: クリック数
        orders:
          type: integer
          description: 注文数
        sales:
          type: number
          description: 売上
        cost:
          type: number
          description: コスト
        cvr:
          type: number
          description: CVR
        acos:
          type: number
          description: ACOS
        target_acos:
          type: number
          description: ターゲットACOS
        score:
          type: number
          description: 優先度スコア
        reason_codes:
          type: array
          items:
            type: string
          description: 理由コード
        reason_detail:
          type: string
          description: 詳細理由
        lifecycle_state:
          type: string
          enum: [LAUNCH_HARD, LAUNCH_SOFT, GROW, HARVEST]
          description: ライフサイクルステート
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, APPLIED]
          description: ステータス
        approved_at:
          type: string
          format: date-time
        approved_by:
          type: string
        rejected_at:
          type: string
          format: date-time
        rejected_by:
          type: string
        rejection_reason:
          type: string
        is_applied:
          type: boolean
        applied_at:
          type: string
          format: date-time
        suggested_at:
          type: string
          format: date-time

    # 実行履歴
...
paths:
...
  /admin/auto-exact-suggestions:
    get:
      summary: AUTO→EXACT昇格候補一覧
      description: AUTO→EXACT昇格候補の一覧を取得
      tags:
        - AUTO to EXACT
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, APPLIED]
          description: ステータスでフィルタ
        - name: asin
          in: query
          schema:
            type: string
          description: ASINでフィルタ
        - name: lifecycleState
          in: query
          schema:
            type: string
            enum: [LAUNCH_HARD, LAUNCH_SOFT, GROW, HARVEST]
          description: ライフサイクルステートでフィルタ
        - name: minScore
          in: query
          schema:
            type: number
          description: 最小スコアでフィルタ
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
          description: 取得件数
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: オフセット
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      suggestions:
                        type: array
                        items:
                          $ref: '#/components/schemas/AutoExactSuggestion'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /admin/auto-exact-suggestions/summary:
    get:
      summary: AUTO→EXACT候補サマリー
      description: ステータス別の集計を取得
      tags:
        - AUTO to EXACT
      parameters:
        - name: asin
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
  /admin/auto-exact-suggestions/top:
    get:
      summary: 高スコア候補Top N
      description: 高スコアのPENDING候補を取得（クイックレビュー用）
      tags:
        - AUTO to EXACT
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: 取得件数
        - name: asin
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
  /admin/auto-exact-suggestions/approve:
    post:
      summary: AUTO→EXACT候補を承認
      description: 指定した候補を承認（PENDING→APPROVED）
      tags:
        - AUTO to EXACT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - suggestionIds
                - approvedBy
              properties:
                suggestionIds:
                  type: array
                  items:
                    type: string
                approvedBy:
                  type: string
            example:
              suggestionIds: ["uuid-1", "uuid-2"]
              approvedBy: "admin@example.com"
      responses:
        '200':
          description: 成功
  /admin/auto-exact-suggestions/reject:
    post:
      summary: AUTO→EXACT候補を却下
      tags:
        - AUTO to EXACT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - suggestionIds
                - rejectedBy
              properties:
                suggestionIds:
                  type: array
                  items:
                    type: string
                rejectedBy:
                  type: string
                reason:
                  type: string
      responses:
        '200':
          description: 成功
  /admin/auto-exact-suggestions/apply-queued:
    post:
      summary: 承認済み候補をAmazon Ads APIに適用
      description: |
        APPROVED状態の候補をAmazon Ads APIに適用します。

        処理内容:
        1. ターゲットMANUALキャンペーンにEXACTキーワードを作成
        2. AUTOキャンペーンにネガティブEXACTを登録（カニバリゼーション防止）

        **注意**: `AUTO_EXACT_APPLY_ENABLED=true` が設定されている必要があります。
      tags:
        - AUTO to EXACT
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dryRun:
                  type: boolean
                  default: false
                maxItems:
                  type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      applyEnabled:
                        type: boolean
                      message:
                        type: string
                      appliedCount:
                        type: integer
                      partialSuccessCount:
                        type: integer
                      failedCount:
                        type: integer
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            suggestionId:
                              type: string
                            searchTerm:
                              type: string
                            keywordCreated:
                              type: boolean
                            keywordId:
                              type: string
                            negativeCreated:
                              type: boolean
                            error:
                              type: string
  /admin/auto-exact-suggestions/{suggestionId}:
    get:
      summary: AUTO→EXACT候補詳細
      tags:
        - AUTO to EXACT
      parameters:
        - name: suggestionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
        '404':
          description: 候補が見つからない

  # 実行履歴
...